+++ src/alifeSim/Simulation/SimulationStatListenerInf.java
@@ -2,5 +2,5 @@ package alifeSim.Simulation;
 
 public interface SimulationStatListenerInf
 {
-	public void simulationStatChanged(long time,long stepNo,int asps);
+	public void simulationStatChanged(long time,int stepNo,int asps);
 }
